PART4-D: Finish exercise 4.20

+++ controllers/BlogController.js
@@ -3,26 +3,16 @@ const jwt = require('jsonwebtoken')
 const Blog = require('../models/Blog')
 const User = require('../models/User')
 
-const getTokenFrom = (request) => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     return response.json(blogs.map(blog => blog.toJSON()))
 })
 
-blogsRouter.post('/', async (request, response, next) => {
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET) 
-    if (!token || !decodedToken.id) {
+blogsRouter.post('/', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET) 
+    if (!request.token || !decodedToken.id) {
         return response.status(401).json({ error: "Missing or invalid token." })
     }
-    
     const user = await User.findById(decodedToken.id)
 
     const newBlog = {
@@ -47,11 +37,24 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    const blog = await Blog.findByIdAndDelete(request.params.id)
-    if (blog) {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!request.token || !decodedToken.id){
+        return response
+          .status(401)
+          .json({ error: 'Missing or invalid token.' })
+    }
+    
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        response.status(400).end()
+    }
+
+    if(blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
         response.status(204).end()
     } else {
-        response.status(400).end()
+        response.status(403).json({ error: 'operation not allowed.'})
     }
 })
